Keep ELLIPSE_ANNOTATOR in main.py; add DIRECTION ellipse helper to motion.
Restore the original main.py annotator layout and remove player.py so new analytics code extends motion.py without relocating legacy demo setup.
